Biblically Accurate Angels
refers to jokes about how spiritual messengers and attendants of God are described in the Christian Bible versus how they are typically portrayed in Western art. In Western art, angels are typically presented as gentle human-like creatures with wings, but descriptions of angels in the Bible paint them as looking far more terrifying, oftentimes with forms unlike anything seen on Earth.